A lot of things, small and large!
It's not dead. I swear! Added a .gitignore that filters out the game
content, almost all the executables, various types of binary data. Added
a lot more functionality to the engine. I think there's a todo list in
here somewhere? I don't know. Here it is, the last year of code.